+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   * Captain hand rotation
   * Clay Apothecary model scale and body part rotations
   * Elder Lemurian toe rotation
+  * Assassin shoulder rotation
   * Model cut-offs persisting when model is used as a skin
 
 * Removed R2API.Language dependency